Validação dos campos do curso - Angular

+++ crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -7,6 +7,7 @@ import { ErrorDialogComponent } from '../../../shared/components/error-dialog/er
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { error } from 'console';
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -35,17 +36,28 @@ export class CoursesComponent {
   onAdd() {
     this.router.navigate(['new'], { relativeTo: this.route });
   }
+  //Editar Curso
   onEdit(course: Course) {
     this.router.navigate(['edit', course.id], { relativeTo: this.route });
   }
+
   onRemove(course: Course) {
-    this.service.remove(course.id).subscribe(
-      () => {
-        this.refresh(), this.openSnackBar('Curso removido com sucesso', 'X');
-      },
-      (error) => this.onError('Erro ao tentar remover o curso')
-    );
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover este curso',
+    });
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.service.remove(course.id).subscribe(
+          () => {
+            this.refresh(),
+              this.openSnackBar('Curso removido com sucesso', 'X');
+          },
+          (error) => this.onError('Erro ao tentar remover o curso')
+        );
+      }
+    });
   }
+
   refresh() {
     this.courses$ = this.service.list().pipe(
       catchError((error) => {
@@ -58,7 +70,7 @@ export class CoursesComponent {
   openSnackBar(message: string, action: string) {
     this.snackBar.open(message, action, {
       duration: 2000,
-      verticalPosition: 'top',
+      verticalPosition: 'bottom',
       horizontalPosition: 'center',
     });
   }
